backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
Pillow==10.2.0
geoalchemy2==0.14.3
python-dotenv==1.0.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
